first attempt at fixing the issue
